Add support for following wiki links in a buffer
This patch adds support for automatically following wiki links in a
buffer. If you are on a link and you hit the enter key or \C-c\C-f the
document pointed to by the wiki link will be opened in another buffer.

This functionality was originially in wiki mode by Alex Schroeder
<alex@gnu.org>. It has been heavily modified and moved here by the
patch author.