Refactor indentation functions to to fix a few issues
When we were at column 0 and the previous line was not indented,
nothing happened.  Also, sometimes the indentation points would
not be unique (e.g., column 0 and indent minus tab-width were
equal), causing us to stop cycling.  This implementation also
lets us make column 0 the default stop when enter is pressed.
Previously, it was skipped over if it was the first element of
the positions list.