Fix matching of pre blocks inside nested lists and at point-min and point-max
The test for (equal (point) cur-end) caused the loop to incorrectly
stop when parsing nested lists.  cur-end would be the same for the
list block as a whole, and for the nested pre block, but cur-begin
would be different for the two, so we should process them
independently.

Also, the inner loop was not executing when a pre block was started at
the beginning of the buffer because prev-indent (which was being
checked) was being initialized before the loop.  Now use a different
indicator variable to determine if the loop executed (and that a
previous block was indeed found).