Bold, italic, link, and image insertion using region or word at point
These commands now use either the region, if there is an active region,
the current word, if the point is at a word, or else insert empty markup.

Cursor positioning is determined according to what should be entered
next.

Bold and italic insertion commands can be used to remove bold and
italic markup when the point is at either a bold or italic phrase.
Cursor position is maintained after the removal.

To support these changes, markdown-wrap-or-insert was rewritten to
accept a thing argument.  In the cases above, thing is 'word.
This function returns the bounds of the region or thing wrapped,
if any, or else it returns nil.