Generalized functions for wrapping, unwrapping, and positioning
Rewrite insertion functions for italics, bold, and code so that they
use the new functions markdown-point-after-unwrap,
markdown-unwrap-thing-at-point, and
markdown-unwrap-things-in-region.

These commands will now wrap the region, if there is an active region.
In this case, they will remove existing markup of the same kind from
the region (since one can't have double italics, for instance).

If there is no active region, but the point is at a word, the markup
will be toggled (removed if it exists, or else added).

Otherwise, empty markup of the specified type will be inserted.

Includes unit tests for markdown-unwrap-things-in-region
and markdown-unwrap-thing-at-point.