Pass over pre blocks when calculating list levels
Previously, markdown-calculate-list-levels and
markdown-update-list-levels were only called from
markdown-match-pre-blocks, so it they did not need to be aware of
lines that matched markdown-regex-list that reside inside pre blocks.
This patch makes them pass over such lines, so that these functions
may be called at an arbitrary point in the buffer to establish the
level and indentation of nested list items.

Includes two unit tests.

Note that these functions currently strictly enforce the four space
rule for nested lists.  Nested item list markers must be positioned
exactly four spaces from the list marker for the parent item, no more
or no less.  This is "safe" for most all Markdown processors, whereas
there is much variation in behavior for other types of spacing.  This
is important because in order for a block to be recognized as a pre
block within a nested list by markdown-mode, it must be indented four
additional spaces for each list level.  For example, in a second-level
list item, a pre block needs to be indented 4 + 2*4 = 12 spaces.