Completion and cycling of headers and horizontal rules
Generic Promotion and Demotion:

Two new functions, markdown-promote and markdown-demote, handle
promotion and demotion of objects at the point.  These are bound
to C-r (right or raise) and C-l (left or lower) respectively.
Currently, these work on headers, horizontal rules, and list
items.  Alternate bindings are M-<right> and M-<left>, which
previously were used only for increasing and decreasing list item
indentation.

Generic Completion and Cycling:

Completion and cycling is carried out using C-c C-=
(markdown-complete-or-cycle).  This function will complete markup
of the object at the point, if it is deemed to be
incomplete (e.g., too few characters underlining a setext-style
header).  Otherwise, it will cycle forward (or backward, with C-u
prefix) through the possible levels.  Currently, this works with
headers and horizontal rules.

Completion and Regions:

Completion with an active region results in all objects in the
region being completed.  Another command, markdown-complete-buffer
(C-c C-c =), has been added for completing all objects in the
buffer.

Headers:

Completion of a header means normalizing the whitespace and
surrounding markup.  Any extraneous whitespace is removed from
the header text, leading and trailing hash marks are added or
equalized (atx headers), and underline characters are extended or
removed to match the text (setext headers).

Horizontal Rules:

The variable markdown-hr-string has been replaced by a list
markdown-hr-strings containing all possible horizontal rule strings to
cycle through.  By default, the first is inserted, but with a C-u
prefix, the last is inserted.  With a numeric prefix C-u n, insert the
n-th element, starting from one.

Completion of a horizontal rule means, check to see if the rule is a
member of markdown-hr-strings.  If so, leave it, otherwise, replace it
with the first string in the list.

Other Changes:

The names of markdown-promote-list-item and markdown-demote-list-item
have been reversed, since promotion actually means reduce the level
and demotion means to increase the level.

Includes unit tests.