Generalize and test markdown-reference-definition function
Rename markdown-has-reference-definition to
markdown-reference-definition, move to parsing section.  This is
general enough to work with footnotes as well.

Instead of returning only the text of the reference definition,
return a list of the form (text start end), where start and end

Leaves match data intact for additional processing.

Includes unit tests.