Improve inline code matching in edge cases
Because the Emacs regular expression does not support negative lookahead
or negative lookbehind, the regular expression for code must include
a character before (to ensure that the backquote is not escaped) and
a character behind (to ensure that another backquote does not follow).
The latter condition is because Markdown allows one to use multiple
backquotes to escape backquotes inside the code fragment.  This means
that font-lock engine in Emacs will mark the leading and trailing
characters as being fontified, preventing them from counting towards
a previous or subsequent inline code fragment.  So, with the old regular
expression, there must be at least two characters between inline code
fragments.  So, `a` `b` `c` would not match properly.

The solution here is to use a matching function which can reposition
the point, so that we move one character back from the start position
and leave the point one character before the last matched by the regular
expression.

Includes test cases.