Improved reference link insertion with label completion
Now uses word at point as link text, if possible, when there is no
active region.  Uses `completing-read' to read reference labels, which
are listed by the new function `markdown-get-defined-references'.

Reference link insertion is smarter now.  It does not prompt for
a URL or title if the label is already defined.  If no URL is given,
create an empty reference definition, but do not prompt for a title.

Includes unit tests for reference link insertion and reference
label listing.

Thanks to Chris Lott for suggesting label completion.