Handle an edge case involving footnote markers and inline links
This patch fixes font lock in cases where a footnote marker precedes
an inline link: Harvard[^1] [tuition][].  The current solution is to
disallow carats as the first character of inline link text.  This is
not ideal, but this is probably more rare than links after footnotes.

Includes a unit test.