Dynamic loading and unloading for math support
Allow `markdown-enable-math' to be set as a local file variable, with
a predicate for ensuring it's safe.

Make `markdown-mode-font-lock-keywords' a local variable.

Add a font lock refresh function `markdown-reload-extensions', which
updates the font-lock keyword list for the current settings and
refontifies the buffer.

Rename regexes with -latex to more generic -math.

Includes unit tests for file local variables, dynamic loading, and
font lock.