On-the-fly toggling for math support, refontification
Add a predicate for making markdown-enable-math safe as a file
variable.

Make `markdown-mode-font-lock-keywords' a local variable.

Add a font lock refresh function markdown-font-lock-refresh-defaults,
which updates the font-lock keyword list for the current settings and
refontifies the buffer.

This restructuring will make it possible to dynamically support
additional extensions in the future.