Customizable reference link location
Same location options as with markdown-footnote-location: 'end, 'header,
or 'immediately.  Default is 'header.

In cases where the point is left after the newly link (when the URL is
given on the minibuffer prompt), print a message to indicate that the
definition was added, since it might not be visible from the link
location.

Includes basic unit tests.