Place definitions at end of 'block'
Now that block-movement functions exist, when location setting is
'immediately, use end of block rather than end of paragraph.  The
current definition of paragraph distinguishes, for example, individual
list items (regardless of whitespace) and hard line breaks.  The
definition of block does not.  Placing definitions at the end of the
block best matches what one intends with 'immediately, and what happened
when this was originally implemented.