Revised behavior of before and after export hooks
In a previous commit, these hooks weere renamed "hook" instead of
"hooks", which is more in line with standard Emacs hook naming.

    markdown-before-export-hook
    markdown-after-export-hook

With this commit, these hooks are no longer given arguments (so they are
now "normal", rather than "abnormal").  The output filename should not
be needed, since modifications made by these hooks to the source and
output buffer will be saved with the exported file.

The before-export hook is now able to modify the buffer, and the buffer
will be restored after export.  Since this is being done before export,
it makes sense to work on the buffer, rather than the output file, which
would be overwritten.  For the after-export hook, any modifications made
to the output buffer will remain and be saved in the exported file.