Fix byte compilation warnings for markdown-test.el
This includes several changes:

1. Removes (unless (featurep 'markdown-mode) ...) from around (require
'markdown-mode). That's redundant because "require" already checks if
the package symbol is in features. But it also breaks things because
require forms are only evaluated by the byte-compiler when they are
top-level forms:
http://www.gnu.org/software/emacs/manual/html_node/elisp/Named-Features.html#Named-Features

2. Adds an explicit (require 'ert) (so it doesn't require autoloads)

3. Replaces several instances of "(goto-line n)" with
"(goto-char (point-min)) (forward-line (1- n))" (as suggested in the
docstring for goto-line.

4. Defines a local function with cl-flet instead of with
a (non-top-level) defun. This adds a dependency on 'cl.

Thanks to Samuel Freilich <sfreilich@google.com> for the patch.