Correct fix for a markdown-tests.el byte-compilation warning, add tests.
bccb000 added "(require 'cl)" to markdown-test.el to fix a byte-compilation
warning, but it should have been "(require 'cl-lib)". This did not get caught
before because cl-lib was being autoloaded anyways, and the change was not
properly covered by automated tests.

This change fixes that, and adds tests for the byte-compilation of
markdown-mode.el and markdown-tests.el, running the tests from the byte-compiled
versions.