Fix and improve ordered list behavior
Some people prefer to leave two spaces after the period in ordered lists
so that the alignment agrees with that of double-digit list items:

1.  List item 1
2.  List item 2
10. List item 10

This patch preserves the spacing of the previous list item but also
adjusts for changes in the length of the numerical prefix.

This also fixes a bug in which `markdown-insert-list-item' entered an
infinite loop when called with the point placed just before the first
ordered list item in the buffer.  Now when cases such as these are
encountered, the function will just use a prefix of "1. ".  For
example,

<point>
1.  One
2.  Two

becomes:

1.  <point>
1.  One
2.  Two

Thanks to Roger Bolsius <roger.bolsius@gmail.com> for a patch on which
this one was based.