Adjust ordered list whitespace when marker digit count increases
For example, when the previous item was number 9, adjust the spacing for
item number 10 if possible to preserve alignment, since "10" is one
character wider than "9". This yields lists like

    9.  List item 9
    10. List item 10

instead of

    9.  List item 9
    10.  List item 10

However, the spacing shouldn't be adjusted if there would be no
remaining whitespace, since list markers must be followed by one or more
spaces or a tab. If tabs are used as whitespace, don't attempt to
adjust the spacing.