Set font style for GFM-style code blocks in base markdown-mode, too.
Github Flavored Markdown isn't the only one that supports fenced-code blocks
annotated with a string controlling things like syntax highlighting. That's also
part of the Common Mark spec (http://spec.commonmark.org/0.20/#info-string).

There's room to be permissive here. The syntax for a GFM-style code block is
unlikely to show up in Markdown code in other contexts. And failing to do so
makes those code-blocks much harder to read in the source (e.g. because
Python comments are highlighted as headers and so on).