handle ikiwiki-style search for wikilinks
ikiwiki (at <http://ikiwiki.info/>) supports searching for wikilink
targets not just relative to the current directory, but also relative to
parent directories (up to the wiki root).

This patch supports that search unconditionally.  It also goes beyond
the wiki root, because the emacs mode has no way of knowing where that
root is.  I hope that I have got the termination logic right but the
platform where it's most likely to be wrong (Windows) is currently
inaccessible to me.