Fix test breakage in test-markdown-subtree/move-up.
The test is broken prior to Emacs 24.5 due to Emacs bug #19102:
https://debbugs.gnu.org/cgi/bugreport.cgi?bug=19102
https://github.com/emacs-mirror/emacs/commit/b3910f

While this isn't a bug in the latest version of Emacs 24, it does impact
the version of Emacs 24 in e.g. the most recent long-term-support
distribution of Ubuntu (14.04; the emacs package for that provides Emacs
24.3).

Stephen Berman notes in the bug report (message #14):

> The error occurs when the sexp `(= (char-after) ?\n)' in
> outline-move-subtree-down is evaluated at eob. This can happen either
> when calling outline-move-subtree-down on either of the last two
> subtrees, or when calling outline-move-subtree-up on the last subtree.

Failed to find an easy way to work around that in
markdown-mode-move-subtree-up, so this does not change that function.
Instead, it adjusts the test to avoid triggering that outline-mode bug.
It also changes the assertion about the test-markdown-subtree/move-up
to assert the contents of the error message instead of asserting that
the type of the error is 'user-error. That's because prior to the change
that fixed this bug in Emacs 24.5, the type of that error message was
simply 'error.