Use text properties to mark code blocks
During the syntax analysis phase, mark pre blocks, tilde fenced code
blocks, and GFM quoted code blocks with text properties. Use these
properties for determining syntax later when setting font-lock faces
and when to determine when the point is within a code block.

As a side effect of these changes, we now no longer require a blank
line preceeding GFM code blocks. This is recommended for style
reasons, but apparently it is not actually required.