More selective wiki link refontification
We only need to refontify wiki links in a region when
the range already contains text with a wiki link face
or if the wiki link regexp matches text in the region.
Otherwise, we can be sure there are no wiki links in
the region and can skip refontifying wiki links
(which is expensive in an after-change hook).