add nuance to markdown-live-preview-delete-export
features like cross-references (#refs) don't work in
markdown-live-preview-mode, because it deletes the output html file on
each export. this can be turned off, but then an html file is left over
after an editing session, and has to be manually deleted. this adds the
ability to only delete the output file automatically after quitting a
markdown-live-preview-mode session.