gfm-mode line wrap behavior and GFM documentation
`visual-line-mode` is no longer enabled by default in `gfm-mode` since
newlines are only significant on GitHub in issues and comments, but not
for actual files in repositories such as README.md, etc.

The documentation for GFM-specific features has been rewritten to
explain in more detail which features are supported by `markdown-mode`
and `gfm-mode`.

Adds a `gfm-mode-hook` for those who want to keep `visual-line-mode` on
by default in `gfm-mode`.

See GH-31 for additional details.