use lexical-binding and markdown-async-result
allow for possibility of process completing in between evaluating links in
sync-do-async chain by creating markdown-async-result. use
lexical-binding so that lambdas have access to lexically let-bound
variables.

lexical-binding is required globally; the alternative is to change every
let(*) which encloses a markdown-do-sync-or-async call to
lexical-let(*), which is difficult to keep track of and inconsistent.