fix for highly confusing tramp error
If you put something like

[[/!foo website="http://bar.baz"]]

in a markdown file, the conversion to filename will start failing
with tramp-related errors.  This is particularly bad because
fontification happens on window configuration change, so killing
the buffer and/or switching away from it also fail.

Fix: for the normal case, escape without testing the filesystem.
Wrap a condition-case around for when the user has specified
search-parent-directories.