Handle false positive italics across list items
Simply checking for `markdown-list-face` in the potential italic range
seems to be a safer and more general approach.  It also covers the
hanging list test case in 6aafece and is more efficient since we've
already done the hard work of parsing the nested lists.

Also, remove an ambiguous test case. With the example below, Markdown.pl
does indeed generate a list item with italics:

* foo_bar
foo_bar