syntax-propertize region should include entire code blocks
This commit ensures that the region for syntactic analysis does not
begin or end in the middle of fenced or indented code blocks.

This fixes a case where with two adjacent code blocks, and with the
point at the empty line in between, when typing text the region would
extend to include only the inner tilde fence beginning with line
5 (`~~~`) and ending with line 7 (`~~~ shell`):

    ~~~ shell
    #!/bin/sh

    echo "Hello, world!"
    ~~~

    ~~~ shell
    #!/bin/sh

    echo "Hello, world v2!"
    ~~~

In other words, these two code blocks would be parsed correctly upon
first loading `markdown-mode`, however, dynamically an incorrect code
block could be identified during syntactic analysis after a change. The
result is that the hash mark in the second code block would be
identified as a heading.