Correct beginning of inline code position
markdown-regex-code starts with "\(?:\`\|[^\]\)\(\(`+\)...". If cursor is
beginning of buffer, "(match-beginning 0)" is on backtick. While cursor is not
beginning of buffer, "(match-beginning 0)" is not on backtick and previous point
of backtick.

---
foo:bar
---
`x`

In above example, "(match-beginning 0)" is point of after "---". That point is
a code block and "(markdown-code-block-at-pos pos)" returns non-nil,
this causes infinite loop of markdown-match-inline-generic.
"(match-beginning 1)" of markdown-regex-code always on backtick. We should
use "(match-beginning 1)" instead of "(match-beginning 0)" in such case.