Add 'C-u C-c C-o' to open links with find-file
I edit a lot of Markdown that cross-links local files. This commit
adds the ability to follow those links with find-file, instead of
opening them in a browser. The fragment-removal in
markdown-link-filename supports links like:

  See [the README](README.md#license).

to work.

It would be nice if the find-file logic used the fragment in such
links to scroll to the appropriate anchor in the target file.  But
often that target is autogenerated from a slugged header, and that
seemed to magical to be worth figuring out.

It would be nice if the find-file vs. browse-url logic happened
automatically, but I'm not fluent enough in Emacs Lisp for that to be
worth the trouble.