Add used language names to front of list
Previously, when a truly new language was used, it was added to the
front of the list as returned by `markdown-gfm-get-corpus`.  However,
when a known language was used it was simply kept in position in the
list of recognized languages. This yielded unexpected results with ido,
helm, etc. Now, we add all used languages to the history list
`markdown-gfm-used-languages`.  This list is then prepended to the list
of recognized languages and returned by `markdown-gfm-get-corpus`.