Unmark buffer modified at markdown-unfontify-region-wiki-links
In Emacs 24.3 some `remove-text-properties` calls, like the ones in
`markdown-unfontify-region-wiki-links`, mark the buffer modified. This
function is invoked by default from `markdown-mode`, and so it makes
that a Markdown buffer is marked modified just by visiting a file or
changing to this major mode.

This change unsets the buffer modified mark if it wasn't previously set.