Prohibit setext headers from starting with spaces or tabs
Fixes an error in the setext header regular expressions. The hyphen
needs to be at the end of the class so that it isn't interpreted as a
range.  This also adds tab as a prohibited character.

This closes GH-143 and is related to GH-139. Thanks to Howard Melman
for reporting these issues.