Use GitHub contributors graph in lieu of manual contribution list
In the early days of markdown-mode, there were far fewer contributors
and the contributions were more specific, so listing them by hand on
the homepage was fairly easy.  However, the existing list has become
outdated and now that contributions are more frequent, it is more
difficult to maintain by hand.  I sincerely appreciate each and every
contribution, but it's probably better for me to spend that time
working on markdown-mode.  Fortunately, we now have an
easily-accessible Git history of the project, so going forward I
will simply link to the GitHub contributors graph available here:
<https://github.com/jrblevin/markdown-mode/graphs/contributors>.