Better fix for compiler warning
Hi,

In mardown-mode you installed

    commit da6f82a260fd647031ddbe5f46fc2f15aa9fd33c
    Author: Jason Blevins <jrblevin@sdf.org>
    Date:   Mon May 15 21:41:47 2017 -0400

        Fix byte compiler warning

but I suggest you fix it by simply swapping the defalias and the if
instead.  This way the byte-compiler doesn't need to collect and compare
the set of definitions present in each branch of the `if` (which it
doesn't bother to do), and instead it immediately sees that
`markdown-directory-files-recursively` is always defined.

Here's a corresponding patch on top of the current head,

        Stefan