Return checkbox string in markdown-cur-list-item-bounds
This makes the function behave as advertised in the docstring.
The "checkbox" element of the return value slipped through
accidentally in the docstring in a previous commit.  This
commit adds that functionality.

The augmented return value is then used in the function
`markdown-insert-list-item`.

Two tests are added regarding GFM task list items, and
existing tests have been updated to reflect the new
functionality.