Refactor markdown-match-generic-metadata to improve efficiency
This is accomplished in two ways.  The first is limiting the scope of
the search to consider only one metadata block (by default, one that
starts at the beginning of the file and ends with the first blank
line).  The second is to be better about leaving the point at an
appropriate position to reduce the amount of times the function is
called by the font-lock engine.

Closes GH-181.