Use a list of possible URL composition characters
With this commit, `markdown-url-compose-char` is permitted to be a
list.  The default value is now a sequence of possible characters and
the first one that is displayable will be used.

This is part of a series of changes to improve the load time by
deferring calls to `char-displayable-p`.  See GH-264.