Use a list for blockquote display characters
With this commit, `markdown-blockquote-display-char` is permitted to
be a list.  The default value is now a sequence of possible strings
and the first one that is displayable will be used.

This is part of a series of changes to improve the load time by
deferring calls to `char-displayable-p`.  See GH-264.