Skip empty URLs when inlining images
Markdown links like `[foo]()`, while basically invalid, do not usually
cause too much trouble, or at least shouldn't cause problems in
unrelated code. It so happens that if a document happens to have such
a link, `markdown-toggle-inline-images` will crash with:

    insert-file-contents-literally: not a regular file: /tmp/default-directory-path/

This patch fixes the problem by checking if the `file` path extracted
from the regex is empty before proceeding with treating it like an image.