Add to beginning of `auto-mode-alist', not end
In Emacs 26, a file called 'CHANGELOG.md' will open in
`changelog-mode' instead of `markdown-mode', because a regexp was
added to `auto-mode-alist' that matches the 'CHANGELOG' part of the
filename. In order for the file to correctly open in `markdown-mode',
the regexps for `markdown-mode' must be added earlier in
`auto-mode-alist', hence this commit.

There should be no concern about this overriding user configuration,
since it makes no sense for user configuration code to be run before
the autoloads for `markdown-mode' are evaluated.