Add option to display images stored online
If non-nil, this option will download images from a remote server (if
the URL is given with a white-listed protocol; by default only
allowing HTTPS) and display them during the execution of
`markdown-display-inline-images'.

Absolute process-specific caching is employed to avoid
double-downloading images.  No `nice' means is yet given for the
invalidation of this cache.