Declare variable aliases before their referent
Since Emacs 27 the byte-compiler tells us to do that.
The warning is shown by `byte-compile-file-form-defvar-function',
which contains a comment providing this justification:

Variable aliases are better declared before the corresponding variable,
since it makes it more likely that only one of the two vars has a value
before the `defvaralias' gets executed, which avoids the need to
merge values.