Avoid matching code fragments as code blocks
Code fragments should be delimited by one backquote on each side, but
some people use two or even three on each side.  Previously if three
were used then we wrongly identified that as the beginning of a code
block.  This resulted in the complete line being treated as syntax,
and hiding syntax therefore caused the complete line to be hidden.
Fix that by making sure no excess backquotes follow on the same line.

Fixes #403.