feat: Add resize inline image without imagemagick
This is a small enahcnement to `markdown-toggle-inline-images`,
adding a feature to resize the inline display of the image without the need
for Imagemagick present in the computer.

It uses the existing variable `markdown-max-image-size` to
determine the size.

Org Mode now is able to do this (see the code in the link below). This feature
is intended to emulate it in Markdown-mode.

https://github.com/bzg/org-mode/blob/abedf386b3f13af2769728755d00c4698ac5d3b0/lisp/org.el#L16344-L16349

      (create-image file-or-data
		    (and (image-type-available-p 'imagemagick)
			 width
			 'imagemagick)
		    remote?
		    :width width))))