WIP: Add support for TOML metadata/frontmatter
This adds support to recognize “+++\n...\n+++” blocks as TOML, similar
to “---” YAML metadata. This is a work in progress because the regular
expressions should probably be in `defconst` variables, and we should
use the same method to highlight YAML metadata which I think would then
allow a lot of other code to be deleted. Also, tests would be nice!