Tidy: (markdown-test-string-mode)
This corrects a few issues:

- Use of UNWIND-PROTECT with IGNORE in the UNWIND-FORMS is essentially
the same as using IGNORE-ERRORS; but such errors probably shouldn't be
ignored in these tests, anyway.

- Using WITH-TEMP-BUFFER makes calling KILL-BUFFER unnecessary.

- Using PROG1 with a spliced list of BODY forms is bogus, because only
the first form's value would be returned, which would be hidden from
callers of this macro; and it was likely used because of calling
KILL-BUFFER, which isn't necessary, anyway.

With these changes, the tests all still pass.