Fix: (markdown-follow-link-at-point) Don't save-excursion
Using `save-excursion' here means that after
`markdown-follow-link-functions' or `markdown--browse-url' may change
the buffer, the buffer gets changed back to the Markdown buffer.
Besides seeming to be undesirable behavior, it breaks the
`test-markdown-link/follow' test.

It was likely added because of the call to `posn-set-point', but it's
unclear why that would make it necessary, because clicking with the
mouse generally moves the point.

If clicking with the mouse really should not move point, this could be
worked around, but `save-excursion' seems generally incompatible with
calling functions which may leave the buffer changed according to the
user's action.