Check if transient-mark-mode is on
This doesn't make sense for all commands like header insertion, so
I only implemented it for some commands. Some of them said in the
documentation that they check if transient-mark-mode was on which
they didn't seem to do but they do with this commit. This is the way
vertico and a couple of other packages use (use-region-p) in a more
legacy-friendly way.