Get rid of some quadratic behavior
markdown-match-bold and markdown-match-italic both call
markdown-inline-code-at-pos-p on each candidate match,
which takes linear time in the distance
from the start of the block to the point being checked.

In a large block with many candidate matches inside inline blocks,
this is slow because the function is called with each candidate, resulting in
overall quadratic behavior, which this commit fixes by starting the
code-at-pos scan after the last already-found match.

For example, the following benchmark code:

(with-temp-buffer
  (dotimes (_ 400)
    (insert "`my_test_string`\n"))
  (markdown-mode)
  (car (benchmark-run
           (font-lock-debug-fontify))))

takes about 10.5 seconds before this commit, and about 0.14 seconds after
this commit.