Don't break inline-code rendering when there's a <foo:bar>
A Markdown like `value=<addr:port>` results in `<addr:port>` suddenly
refusing to highlight. This is because the mode recognizes this as an
angle URL. Generally speaking, a URL (even if actual one) is not
expected to be rendered inside inline code-blocks, so this commit adds
a test for whether start or end is in inline block, to avoid breaking
rendering.

A test for this is added as well.

Fixes: https://github.com/jrblevin/markdown-mode/issues/822